Move apiclient_util and apiclient_util_test to apiclient/util package
